+++ src/main/java/brig/concord/psi/ProcessDefinitionProvider.java
@@ -1,9 +1,6 @@
 package brig.concord.psi;
 
-import com.intellij.openapi.vfs.VirtualFile;
 import com.intellij.psi.PsiElement;
-import com.intellij.psi.PsiFile;
-import com.intellij.psi.PsiManager;
 import com.intellij.util.AstLoadingFilter;
 import org.jetbrains.yaml.psi.YAMLDocument;
 
@@ -24,21 +21,23 @@ public class ProcessDefinitionProvider {
         if (currentDoc == null) {
             return null;
         }
+        return new ProcessDefinition(currentDoc);
 
-        YAMLDocument rootDoc;
-
-        VirtualFile rootFile = YamlPsiUtils.rootConcordYaml(element);
-        if (rootFile != null) {
-            PsiFile rootPsiFile = PsiManager.getInstance(element.getProject()).findFile(rootFile);
-            rootDoc = YamlPsiUtils.getDocument(rootPsiFile);
-        } else {
-            rootDoc = currentDoc;
-        }
-
-        if (rootDoc == null || rootDoc.getContainingFile() == null || rootDoc.getContainingFile().getVirtualFile() == null) {
-            return null;
-        }
-
-        return new ProcessDefinition(rootDoc);
+// TODO: find root, process resources configuration...
+//        YAMLDocument rootDoc;
+//
+//        VirtualFile rootFile = YamlPsiUtils.rootConcordYaml(element);
+//        if (rootFile != null) {
+//            PsiFile rootPsiFile = PsiManager.getInstance(element.getProject()).findFile(rootFile);
+//            rootDoc = YamlPsiUtils.getDocument(rootPsiFile);
+//        } else {
+//            rootDoc = currentDoc;
+//        }
+//
+//        if (rootDoc == null || rootDoc.getContainingFile() == null || rootDoc.getContainingFile().getVirtualFile() == null) {
+//            return null;
+//        }
+//
+//        return new ProcessDefinition(rootDoc);
     }
 }